pyproject.toml
[tool.poetry]
name = "voidmon"
version = "0.1.0"
description = ""
authors = ["Tristan Kernan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-extensions = "^3.2.3"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"
django-htmx = "^1.21.0"
gunicorn = "^23.0.0"
python-decouple = "^3.8"
whitenoise = "^6.8.2"
django-tables2 = "^2.7.4"
django-filter = "^24.3"
redis = "^5.2.1"
sentry-sdk = "^2.19.2"
celery = "^5.4.0"
requests = "^2.32.3"
django-celery-results = "^2.5.1"
django-allauth = "^65.3.1"
django-anymail = "^12.0"
werkzeug = "^3.1.3"
factory-boy = "^3.3.1"
dj-database-url = "^2.3.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
ipdb = "^0.13.13"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
time-machine = "^2.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
